+++ front/components/poke/pages/ConversationPage.tsx
@@ -1,13 +1,12 @@
 import { useDocumentTitle } from "@app/hooks/useDocumentTitle";
-import { useSendNotification } from "@app/hooks/useNotification";
 import { useWorkspace } from "@app/lib/auth/AuthContext";
 import { clientFetch } from "@app/lib/egress/client";
 import { useRequiredPathParam } from "@app/lib/platform";
 import { classNames } from "@app/lib/utils";
-import type { GetReinforcementTestCaseResponseBody } from "@app/pages/api/poke/workspaces/[wId]/conversations/[cId]/reinforcement_test_case";
 import { usePokeConversation } from "@app/poke/swr";
 import { usePokeAgentConfigurations } from "@app/poke/swr/agent_configurations";
 import { usePokeConversationConfig } from "@app/poke/swr/conversation_config";
+import { useCopyReinforcementTestCase } from "@app/poke/swr/reinforcement_test_case";
 import type {
   CompactionMessageType,
   UserMessageType,
@@ -387,37 +386,8 @@ export function ConversationPage() {
   const [showRenderControls, setShowRenderControls] = useState(false);
   const [isCopiedJSON, copyJSON] = useCopyToClipboard();
 
-  const [isTestCaseLoading, setIsTestCaseLoading] = useState(false);
-  const sendNotification = useSendNotification();
-
-  async function handleCopyReinforcementTestCase() {
-    setIsTestCaseLoading(true);
-    try {
-      const response = await clientFetch(
-        `/api/poke/workspaces/${owner.sId}/conversations/${conversationId}/reinforcement_test_case`
-      );
-      const data = await response.json();
-      if (!response.ok) {
-        // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
-        throw new Error(data.error?.message || "Failed to generate test case");
-      }
-      const { testCase } = data as GetReinforcementTestCaseResponseBody;
-      await navigator.clipboard.writeText(JSON.stringify(testCase, null, 2));
-      sendNotification({
-        title: "Copied",
-        description: "Reinforcement test case copied to clipboard.",
-        type: "success",
-      });
-    } catch (e) {
-      sendNotification({
-        title: "Error",
-        description: e instanceof Error ? e.message : "Unknown error",
-        type: "error",
-      });
-    } finally {
-      setIsTestCaseLoading(false);
-    }
-  }
+  const { copyTestCase, isLoading: isTestCaseLoading } =
+    useCopyReinforcementTestCase({ owner, conversationId });
 
   useEffect(() => {
     if (!selectedAgentId) {
@@ -563,7 +533,7 @@ export function ConversationPage() {
               label="Reinforcement test"
               variant="primary"
               size="xs"
-              onClick={() => void handleCopyReinforcementTestCase()}
+              onClick={() => void copyTestCase()}
               disabled={isTestCaseLoading}
             />
             {isRendering && <Spinner size="xs" />}
